+++ contracts/tests/ccip/offramp/OffRamp.Setup.ts
@@ -17,7 +17,7 @@ import { contractCode } from '../../../wrappers/codeLoader'
 import * as ocr from '../../../wrappers/libraries/ocr/MultiOCR3Base'
 import * as OCR3Logs from '../../../wrappers/libraries/ocr/Logs'
 import * as deployable from '../../../wrappers/libraries/Deployable'
-import { PERMISSIONLESS_EXECUTION_THRESHOLD_SECONDS } from './OffRamp.commitAndExec.spec'
+import { PERMISSIONLESS_EXECUTION_THRESHOLD_SECONDS } from './OffRamp.execute.spec'
 import { ChainSelectors } from '../../utils/Selectors'
 import { setupTestFeeQuoter } from '../helpers/SetUp'
